chore(suite-native): No send, refactoring
