+++ suite-native/app/src/navigation/RootStackNavigator.tsx
@@ -13,7 +13,7 @@ import {
 import { selectIsOnboardingFinished } from '@suite-native/module-settings';
 import { DevUtilsStackNavigator } from '@suite-native/module-dev-utils';
 import { TransactionDetailScreen } from '@suite-native/transactions';
-import { SendReceiveBottomSheet } from '@suite-native/module-send-receive';
+import { ReceiveModalScreen } from '@suite-native/module-send-receive';
 
 import { AppTabNavigator } from './AppTabNavigator';
 
@@ -52,7 +52,7 @@ export const RootStackNavigator = () => {
             <RootStack.Group screenOptions={{ presentation: 'modal' }}>
                 <RootStack.Screen
                     name={RootStackRoutes.SendReceive}
-                    component={SendReceiveBottomSheet}
+                    component={ReceiveModalScreen}
                 />
             </RootStack.Group>
         </RootStack.Navigator>
