+++ suite-native/module-send-receive/src/index.ts
@@ -1,1 +1,1 @@
-export * from './components/SendReceiveBottomSheet';
+export * from './screens/ReceiveModalScreen';
